+++ src/main/kotlin/com/knu/mockin/model/dto/kisresponse/quotations/basic/mock/KISDailyChartPriceResponseDto.kt
@@ -1,13 +1,13 @@
-package com.knu.mockin.model.dto.kisresponse.basic
+package com.knu.mockin.model.dto.kisresponse.quotations.basic.mock
 
 import com.fasterxml.jackson.annotation.JsonProperty
 
 data class KISDailyChartPriceResponseDto (
-        @JsonProperty("rt_cd") val successFailureStatus: String,    // 성공 실패 여부
-        @JsonProperty("msg_cd") val responseCode: String,           // 응답코드
-        @JsonProperty("msg1") val responseMessage: String,          // 응답메세지
-        @JsonProperty("output1") val output1: DailyChartPriceOutput1?,               // 응답상세
-        @JsonProperty("output2") val output2: List<DailyChartPriceOutput2>?,
+    @JsonProperty("rt_cd") val successFailureStatus: String,    // 성공 실패 여부
+    @JsonProperty("msg_cd") val responseCode: String,           // 응답코드
+    @JsonProperty("msg1") val responseMessage: String,          // 응답메세지
+    @JsonProperty("output1") val output1: DailyChartPriceOutput1?,               // 응답상세
+    @JsonProperty("output2") val output2: List<DailyChartPriceOutput2>?,
 )
 
 data class DailyChartPriceOutput1 (
